refactor: command line options
Introduce the mechansim for command line option seeding by entrances

The commit builds on https://github.com/hamlet-io/engine/pull/1549 and
starts the transition to sourcing command line options from dynamic
input state.

It defines the pattern by which entrances can provide any command
line option seeding prior to the incoation of the entrance itself. This
validation of the options prior to the entrnace being invoked.
The `deployment` entrance is converted to this pattern. A subsequent
change will convert the remainder of the entrances once the pattern
has been validated.

It also ensures the global commandLineOption variable is kept up to
date even if code is transitioned to use an input seeder rather than
directly calling `addCommandLineOption`.